♻️ refactor : slug 검증 로직 분리  #199
